Improved cucumber rails server

+++ features/support/rails_server.rb
@@ -1,10 +1,14 @@
+# Starts a Rails application server in a fork and waits for it to be responsive
 class RailsServer
+  HOST = 'localhost'.freeze
+
   class << self
     attr_accessor :instance
   end
 
   def self.start(port = nil)
     self.instance = new(port)
+    self.instance.start
   end
 
   def self.stop
@@ -18,27 +22,65 @@ class RailsServer
 
   def initialize(port)
     @port = (port || 3001).to_i
-    @output = StringIO.new("")
-    @pid = fork do
-      unless ENV['DEBUG']
-        $stdout = self.output
-        $stderr = self.output
-      end
-      require 'config/environment'
-      app = ActionController::Dispatcher.new
-      Rack::Handler::Thin.run(app, :Port => @port, :AccessLog => [])
-    end
-    sleep(5) # wait for app server to start
   end
 
-  attr_accessor :output
+  def start
+    @pid = fork do
+      unless ENV['DEBUG']
+        output = StringIO.new("")
+        $stdout = output
+        $stderr = output
+      end
+      require 'config/environment'
+      app = Identify.new(ActionController::Dispatcher.new)
+      Rack::Handler::Thin.run(app, :Port => @port, :AccessLog => [])
+    end
+    wait_until_responsive
+  end
 
   def stop
-    Process.kill('INT', @pid)
-    Process.wait(@pid)
+    if @pid
+      Process.kill('INT', @pid)
+      Process.wait(@pid)
+      @pid = nil
+    end
   end
 
   def get(path)
-    Net::HTTP.get(URI.parse("http://localhost:#{@port}").merge(path))
+    Net::HTTP.get(URI.parse("http://#{HOST}:#{@port}").merge(path))
+  end
+
+  def wait_until_responsive
+    20.times do
+      if responsive?
+        return true
+      else
+        sleep(0.5)
+      end
+    end
+    raise "Couldn't connect to Rails application server at #{HOST}:#{@port}"
+  end
+
+  def responsive?
+    response = Net::HTTP.start(HOST, @port) { |http| http.get('/__identify__') }
+    response.is_a?(Net::HTTPSuccess)
+  rescue Errno::ECONNREFUSED, Errno::EBADF
+    return false
+  end
+
+  # From Capybara::Server
+
+  class Identify
+    def initialize(app)
+      @app = app
+    end
+
+    def call(env)
+      if env["PATH_INFO"] == "/__identify__"
+        [200, {}, 'OK']
+      else
+        @app.call(env)
+      end
+    end
   end
 end
